requirements.txt
altair==4.0.1
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
black==19.10b0
bleach==3.1.1
blinker==1.4
bokeh==2.0.0
boto3==1.12.17
botocore==1.15.17
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
flake8==3.7.9
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pydeck==0.3.0b2
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
streamlit==0.56.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.2
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
